Always use Long instead Integer.

+++ modules/xfmg/oas/sharedlib/xyna-openapi/generators/xmom-client/src/main/java/com/gip/xyna/openapi/codegen/XmomClientGenerator.java
@@ -222,7 +222,7 @@ public class XmomClientGenerator extends DefaultCodegen {
 
     typeMapping.clear();
     typeMapping.put("boolean", "Boolean");
-    typeMapping.put("integer", "Integer");
+    typeMapping.put("integer", "Long");
     typeMapping.put("long", "Long");
     typeMapping.put("double", "Double");
     typeMapping.put("float", "Double");
